requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.53
python-dotenv==1.0.1
